+++ src/test/tests.ts
@@ -30,7 +30,7 @@ import {
   GRAPHQL_SUBSCRIPTIONS,
 } from '../protocol';
 
-import { createServer, IncomingMessage, ServerResponse } from 'http';
+import { createServer, IncomingMessage, ServerResponse, Server } from 'http';
 import { SubscriptionServer, ExecutionParams } from '../server';
 import { SubscriptionClient } from '../client';
 import { OperationMessage } from '../server';
@@ -198,8 +198,9 @@ const onConnectErrorOptions = {
   schema,
   subscribe,
   execute,
+  isLegacy: true,
   onConnect: (msg: any, connection: any, connectionContext: any) => {
-    connectionContext.isLegacy = true;
+    connectionContext.isLegacy = onConnectErrorOptions.isLegacy;
     throw new Error('Error');
   },
 };
@@ -875,6 +876,9 @@ describe('Client', function () {
     let client: SubscriptionClient;
     let originalClient: any;
     wsServer.on('connection', (connection: WebSocket) => {
+      connection.on('error', (error) => {
+        // ignored for testing
+      });
       connections += 1;
       if (connections === 1) {
         originalClient.close();
@@ -942,6 +946,26 @@ describe('Client', function () {
         done();
       }).to.throw();
     }, 1000);
+  });
+
+  it('should emit event when an websocket error occurs', function (done) {
+    const client = new SubscriptionClient(`ws://localhost:${ERROR_TEST_PORT}/`);
+
+    client.request({
+      query: `subscription useInfo{
+        invalid
+      }`,
+      variables: {},
+    }).subscribe({
+      next: () => {
+        assert(false);
+      },
+    });
+
+    client.onError((err: Error) => {
+      expect(err.message).to.be.equal(`connect ECONNREFUSED 127.0.0.1:${ERROR_TEST_PORT}`);
+      done();
+    });
   });
 
   it('should stop trying to reconnect to the server', function (done) {
@@ -1179,16 +1203,51 @@ describe('Client', function () {
       done();
     }, 500);
   });
+
+  it('should close the connection after inactivityTimeout and zero active subscriptions', function (done) {
+    const subscriptionsClient = new SubscriptionClient(`ws://localhost:${RAW_TEST_PORT}/`, {
+      inactivityTimeout: 100,
+    });
+    const sub = subscriptionsClient.request({
+      query: `subscription useInfo($id: String) {
+        user(id: $id) {
+          id
+          name
+        }
+      }`,
+      operationName: 'useInfo',
+      variables: {
+        id: 3,
+      },
+    }).subscribe({});
+
+    setTimeout(() => {
+      expect(Object.keys(subscriptionsClient.operations).length).to.be.equal(1);
+      sub.unsubscribe();
+      setTimeout(() => {
+        expect(Object.keys(subscriptionsClient.operations).length).to.be.equal(0);
+        setTimeout(() => {
+          expect(subscriptionsClient.status).to.be.equal(WebSocket.CLOSED);
+          done();
+        }, 101);
+      }, 50);
+    }, 50);
+  });
 });
 
 describe('Server', function () {
   let onOperationSpy: any;
+  let server: Server;
 
   beforeEach(() => {
     onOperationSpy = sinon.spy(handlers, 'onOperation');
   });
 
   afterEach(() => {
+    if (server) {
+      server.close();
+    }
+
     if (onOperationSpy) {
       onOperationSpy.restore();
     }
@@ -1226,7 +1285,7 @@ describe('Server', function () {
   });
 
   it('should accept execute method than returns a Promise (original execute)', (done) => {
-    const server = createServer(notFoundRequestListener);
+    server = createServer(notFoundRequestListener);
     server.listen(SERVER_EXECUTOR_TESTS_PORT);
     let msgCnt = 0;
 
@@ -1257,7 +1316,6 @@ describe('Server', function () {
         },
         complete: () => {
           expect(msgCnt).to.equals(1);
-          server.close();
           done();
         },
       });
@@ -1265,7 +1323,7 @@ describe('Server', function () {
   });
 
   it('server close should work', (done) => {
-    const server = createServer(notFoundRequestListener);
+    server = createServer(notFoundRequestListener);
     server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     const subServer = SubscriptionServer.create({
@@ -1278,7 +1336,6 @@ describe('Server', function () {
 
     const client = new SubscriptionClient(`ws://localhost:${SERVER_EXECUTOR_TESTS_PORT}/`);
     client.onDisconnected(() => {
-      server.close();
       done();
     });
 
@@ -1300,7 +1357,7 @@ describe('Server', function () {
   });
 
   it('should have request interface (apollo client 2.0)', (done) => {
-    const server = createServer(notFoundRequestListener);
+    server = createServer(notFoundRequestListener);
     server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
@@ -1325,13 +1382,9 @@ describe('Server', function () {
           hasValue = true;
         },
         error: (err) => {
-          server.close();
-
           done(new Error('unexpected error from subscribe'));
         },
         complete: () => {
-          server.close();
-
           if ( false === hasValue ) {
             return done(new Error('No value recived from observable'));
           }
@@ -1341,36 +1394,8 @@ describe('Server', function () {
     });
   });
 
-  it('should return an error when invalid execute method provided', (done) => {
-    const server = createServer(notFoundRequestListener);
-    server.listen(SERVER_EXECUTOR_TESTS_PORT);
-
-    SubscriptionServer.create({
-      schema,
-      execute: (() => ({})) as any,
-    }, {
-      server,
-      path: '/',
-    });
-
-    const client = new SubscriptionClient(`ws://localhost:${SERVER_EXECUTOR_TESTS_PORT}/`);
-    client.onConnected(() => {
-      client.request({
-        query: `query { testString }`,
-        variables: {},
-      }).subscribe({
-        error: (err) => {
-          expect(err.message).to.equal('GraphQL execute engine is not available');
-          client.client.close();
-          server.close();
-          done();
-        },
-      });
-    });
-  });
-
   it('should accept execute method than returns an AsyncIterator', (done) => {
-    const server = createServer(notFoundRequestListener);
+    server = createServer(notFoundRequestListener);
     server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     const executeWithAsyncIterable = () => {
@@ -1418,8 +1443,6 @@ describe('Server', function () {
           } else {
             expect(res.data).to.deep.equal({ testString: 'value' });
           }
-
-          server.close();
           done();
         },
       });
@@ -1477,9 +1500,36 @@ describe('Server', function () {
     }, 200);
   });
 
+  it('should trigger onConnect with the request available in ConnectionContext', (done) => {
+    new SubscriptionClient(`ws://localhost:${EVENTS_TEST_PORT}/`);
+
+    setTimeout(() => {
+      assert(eventsOptions.onConnect.calledOnce);
+      expect(eventsOptions.onConnect.getCall(0).args[2].request).to.be.an.instanceof(IncomingMessage);
+      done();
+    }, 200);
+  });
+
   it('should trigger onConnect and return GQL_CONNECTION_ERROR with error', (done) => {
     const connectionCallbackSpy = sinon.spy();
 
+    onConnectErrorOptions.isLegacy = false;
+    const subscriptionsClient = new SubscriptionClient(`ws://localhost:${ONCONNECT_ERROR_TEST_PORT}/`, {
+      connectionCallback: connectionCallbackSpy,
+    });
+
+    setTimeout(() => {
+      expect(connectionCallbackSpy.calledOnce).to.be.true;
+      expect(connectionCallbackSpy.getCall(0).args[0]).to.eql({ message: 'Error' });
+      subscriptionsClient.close();
+      done();
+    }, 200);
+  });
+
+  it('should trigger onConnect and return INIT_FAIL with error', (done) => {
+    const connectionCallbackSpy = sinon.spy();
+
+    onConnectErrorOptions.isLegacy = true;
     const subscriptionsClient = new SubscriptionClient(`ws://localhost:${ONCONNECT_ERROR_TEST_PORT}/`, {
       connectionCallback: connectionCallbackSpy,
     });
@@ -1487,18 +1537,21 @@ describe('Server', function () {
     const originalOnMessage = subscriptionsClient.client.onmessage;
     subscriptionsClient.client.onmessage = (dataReceived: any) => {
       let messageData = JSON.parse(dataReceived.data);
-
+      // Reformat message to avoid unknown message type
       if (messageData.type === MessageTypes.INIT_FAIL) {
         messageData.type = MessageTypes.GQL_CONNECTION_ERROR;
       }
-
       dataReceived.data = JSON.stringify(messageData);
       originalOnMessage(dataReceived);
     };
 
     setTimeout(() => {
       expect(connectionCallbackSpy.calledOnce).to.be.true;
-      expect(connectionCallbackSpy.getCall(0).args[0]).to.equal('Error');
+      // Old client used: connectionCallback(parsedMessage.payload.error)
+      // But new client uses: connectionCallback(parsedMessage.payload)
+      // So check complete payload
+      expect(connectionCallbackSpy.getCall(0).args[0]).to.eql({ error: 'Error' });
+      subscriptionsClient.close();
       done();
     }, 200);
   });
@@ -1510,6 +1563,18 @@ describe('Server', function () {
     }, 100);
     setTimeout(() => {
       assert(eventsOptions.onDisconnect.calledOnce);
+      done();
+    }, 200);
+  });
+
+  it('should trigger onDisconnect with ConnectionContext as second argument', (done) => {
+    const client = new SubscriptionClient(`ws://localhost:${EVENTS_TEST_PORT}/`);
+    setTimeout(() => {
+      client.client.close();
+    }, 100);
+    setTimeout(() => {
+      assert(eventsOptions.onDisconnect.calledOnce);
+      expect(eventsOptions.onConnect.getCall(0).args[1]).to.not.be.undefined;
       done();
     }, 200);
   });
@@ -1941,11 +2006,38 @@ describe('Server', function () {
       const parsedMessage = JSON.parse(message.data);
       if (parsedMessage.type === MessageTypes.GQL_CONNECTION_KEEP_ALIVE) {
         yieldCount += 1;
-        if (yieldCount > 1) {
+        if (yieldCount > 2) {
           client.close();
           done();
         }
       }
+    };
+    client.onopen = () => {
+      client.send(JSON.stringify({
+        id: 1,
+        type: MessageTypes.GQL_CONNECTION_INIT,
+      }));
+    };
+  });
+
+  it('sends legacy keep alive signal in the socket', function (done) {
+    let client = new WebSocket(`ws://localhost:${KEEP_ALIVE_TEST_PORT}/`, GRAPHQL_SUBSCRIPTIONS);
+    let yieldCount = 0;
+    client.onmessage = (message: any) => {
+      const parsedMessage = JSON.parse(message.data);
+      if (parsedMessage.type === MessageTypes.KEEP_ALIVE) {
+        yieldCount += 1;
+        if (yieldCount > 2) {
+          client.close();
+          done();
+        }
+      }
+    };
+    client.onopen = () => {
+      client.send(JSON.stringify({
+        id: 1,
+        type: MessageTypes.INIT,
+      }));
     };
   });
 });
@@ -1960,16 +2052,74 @@ describe('Message Types', function () {
 });
 
 describe('Client<->Server Flow', () => {
-  it('should reconnect after manually closing the connection and then resubscribing', (done) => {
-    const testServer = createServer(notFoundRequestListener);
-    testServer.listen(SERVER_EXECUTOR_TESTS_PORT);
+  let server: Server;
+
+  afterEach(() => {
+    if (server) {
+      server.close();
+    }
+  });
+
+  it('should reconnect after inactivityTimeout closing the connection and then resubscribing', (done) => {
+    server = createServer(notFoundRequestListener);
+    server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
       schema: subscriptionsSchema,
       execute,
       subscribe,
     }, {
-      server: testServer,
+      server,
+      path: '/',
+    });
+
+    const client = new SubscriptionClient(`ws://localhost:${SERVER_EXECUTOR_TESTS_PORT}/`, {
+      inactivityTimeout: 100,
+    });
+    let isFirstTime = true;
+
+    client.onConnected(async () => {
+      // Manually close the connection only in the first time, to avoid infinite loop
+      if (isFirstTime) {
+        isFirstTime = false;
+
+        setTimeout(() => {
+          const sub1 = client.request({
+            query: `query { testString }`,
+            variables: {},
+          }).subscribe({});
+          setTimeout(() => {
+            sub1.unsubscribe();
+            setTimeout(() => {
+              const sub2 = client.request({
+                query: `query { testString }`,
+                variables: {},
+              }).subscribe({
+                next: (res) => {
+                  expect(sub2).not.to.eq(null);
+                  expect(res.errors).to.equals(undefined);
+                  expect(res.data.testString).to.eq('value');
+                  sub2.unsubscribe();
+                  done();
+                },
+              });
+            }, 200);
+          }, 50);
+        }, 50);
+      }
+    });
+  });
+
+  it('should reconnect after manually closing the connection and then resubscribing', (done) => {
+    server = createServer(notFoundRequestListener);
+    server.listen(SERVER_EXECUTOR_TESTS_PORT);
+
+    SubscriptionServer.create({
+      schema: subscriptionsSchema,
+      execute,
+      subscribe,
+    }, {
+      server,
       path: '/',
     });
 
@@ -1996,7 +2146,6 @@ describe('Client<->Server Flow', () => {
               expect(res.data.testString).to.eq('value');
 
               sub.unsubscribe();
-              testServer.close();
               done();
             },
           });
@@ -2006,8 +2155,8 @@ describe('Client<->Server Flow', () => {
   });
 
   it('validate requests against schema', (done) => {
-    const testServer = createServer(notFoundRequestListener);
-    testServer.listen(SERVER_EXECUTOR_TESTS_PORT);
+    server = createServer(notFoundRequestListener);
+    server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
       schema: subscriptionsSchema,
@@ -2015,7 +2164,7 @@ describe('Client<->Server Flow', () => {
       subscribe,
       validationRules: specifiedRules,
     }, {
-      server: testServer,
+      server,
       path: '/',
     });
 
@@ -2045,7 +2194,6 @@ describe('Client<->Server Flow', () => {
               );
 
               sub.unsubscribe();
-              testServer.close();
               done();
             },
           });
@@ -2058,15 +2206,15 @@ describe('Client<->Server Flow', () => {
     subscriptionAsyncIteratorSpy.reset();
     resolveAsyncIteratorSpy.reset();
 
-    const testServer = createServer(notFoundRequestListener);
-    testServer.listen(SERVER_EXECUTOR_TESTS_PORT);
+    server = createServer(notFoundRequestListener);
+    server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
       schema: subscriptionsSchema,
       execute,
       subscribe,
     }, {
-      server: testServer,
+      server,
       path: '/',
     });
 
@@ -2135,21 +2283,20 @@ describe('Client<->Server Flow', () => {
     // should be 0 because subscribe called only in the beginning
     expect(subscriptionAsyncIteratorSpy.callCount).to.eq(0);
     client2.unsubscribe();
-    testServer.close();
   });
 
   it('should close iteration over AsyncIterator when client disconnects', async () => {
     resolveAsyncIteratorSpy.reset();
 
-    const testServer = createServer(notFoundRequestListener);
-    testServer.listen(SERVER_EXECUTOR_TESTS_PORT);
+    server = createServer(notFoundRequestListener);
+    server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
       schema: subscriptionsSchema,
       execute,
       subscribe,
     }, {
-      server: testServer,
+      server,
       path: '/',
     });
 
@@ -2212,7 +2359,6 @@ describe('Client<->Server Flow', () => {
     expect(client2.spy.callCount).to.eq(1);
     // should be 1 because there is only one subscriber now (client2)
     expect(resolveAsyncIteratorSpy.callCount).to.eq(1);
-    testServer.close();
   });
 
   it('should handle correctly multiple subscriptions one after each other', (done) => {
@@ -2221,7 +2367,7 @@ describe('Client<->Server Flow', () => {
     // change quickly and we want to make sure that the subscriptions flow is correct
 
     // Create the server
-    const server = createServer(notFoundRequestListener);
+    server = createServer(notFoundRequestListener);
     server.listen(SERVER_EXECUTOR_TESTS_PORT);
 
     SubscriptionServer.create({
@@ -2275,12 +2421,11 @@ describe('Client<->Server Flow', () => {
                 }).subscribe({
                   next: (s2Res) => {
                     assert(s2Res.errors === undefined, 'unexpected error from 2nd subscription');
-                    assert(s2Res.data !== null, 'unexpected null from 2nd ubscription result');
+                    assert(s2Res.data !== null, 'unexpected null from 2nd subscription result');
                     expect(s2Res.data.user.id).to.eq('1');
                     expect(Object.keys(client['operations']).length).to.eq(1);
                     expect(firstSubscriptionSpy.callCount).to.eq(1);
 
-                    server.close();
                     done();
                   },
                 });